requirements.txt
beautifulsoup4
playwright
pyyaml